rename LoadingMode into InitMode. InitModes knows only to modes 'background' and 'fail_fast'. Moved Initialization completley into AsyncApiInitApplicationListener. Fixed some tests.
Took 47 minutes
